Fixup branch to work with the new dual exporter options
