+++ src/main/kotlin/nl/avisi/structurizr/site/generatr/site/PlantUmlExporter.kt
@@ -5,6 +5,7 @@ import com.structurizr.export.Diagram
 import com.structurizr.export.IndentingWriter
 import com.structurizr.export.plantuml.C4PlantUMLExporter
 import com.structurizr.export.plantuml.StructurizrPlantUMLExporter
+import com.structurizr.model.Component
 import com.structurizr.model.Container
 import com.structurizr.model.Element
 import com.structurizr.model.SoftwareSystem
@@ -78,9 +79,10 @@ private class WriterWithElementLinks(
         writeElementFn: (view: ModelView?, element: Element?, writer: IndentingWriter?) -> Unit
     ) {
         val url = when {
-            needsLinkToSoftwareSystem(element, view) -> getUrlToSoftwareSystem(element)
-            needsLinkToContainerViews(element, view) -> getUrlToContainerViews(element)
-            needsLinkToComponentViews(element, view) -> getUrlToComponentViews(element)
+            needsLinkToSoftwareSystem(element, view) -> getUrlToViewsPage(element?.name?.normalize(), "context")
+            needsLinkToContainerViews(element, view, workspace) -> getUrlToViewsPage(element?.name?.normalize(), "container")
+            needsLinkToComponentViews(element, view, workspace) -> getUrlToViewsPage(element?.parent?.name?.normalize(), "component")
+            needsLinkToCodeViews(element, workspace) -> getUrlToViewsPage(element?.parent?.parent?.name?.normalize(), "code")
             else -> null
         }
 
@@ -93,24 +95,23 @@ private class WriterWithElementLinks(
     private fun needsLinkToSoftwareSystem(element: Element?, view: ModelView?) =
         element is SoftwareSystem && workspace.includedSoftwareSystems.contains(element) && element != view?.softwareSystem
 
-    private fun getUrlToSoftwareSystem(element: Element?): String {
-        val path = "/${element?.name?.normalize()}/context/".asUrlToDirectory(url)
-        return "$TEMP_URI$path"
-    }
+    private fun needsLinkToContainerViews(element: Element?, view: ModelView?, workspace: Workspace) =
+        element is SoftwareSystem
+                && workspace.includedSoftwareSystems.contains(element)
+                && element == view?.softwareSystem
+                && ( element.hasContainers || getImageViewsForId(workspace, element.id).isNotEmpty())
 
-    private fun needsLinkToContainerViews(element: Element?, view: ModelView?) =
-        element is SoftwareSystem && workspace.includedSoftwareSystems.contains(element) && element == view?.softwareSystem && element.hasContainers
+    private fun needsLinkToComponentViews(element: Element?, view: ModelView?, workspace: Workspace) =
+        element is Container
+                && ( element.hasComponents || getImageViewsForId(workspace, element.id).isNotEmpty())
+                && view !is ComponentView
 
-    private fun getUrlToContainerViews(element: Element?): String {
-        val path = "/${element?.name?.normalize()}/container/".asUrlToDirectory(url)
-        return "$TEMP_URI$path"
-    }
+    private fun needsLinkToCodeViews(element: Element?, workspace: Workspace) =
+        element is Component
+                && getImageViewsForId(workspace, element.id).isNotEmpty()
 
-    private fun needsLinkToComponentViews(element: Element?, view: ModelView?) =
-        element is Container && element.hasComponents && view !is ComponentView
-
-    private fun getUrlToComponentViews(element: Element?): String {
-        val path = "/${element?.parent?.name?.normalize()}/component/".asUrlToDirectory(url)
+    private fun getUrlToViewsPage(systemName: String?, page: String): String {
+        val path = "/$systemName/$page/".asUrlToDirectory(url)
         return "$TEMP_URI$path"
     }
 
@@ -147,7 +148,7 @@ class C4PlantUmlExporterWithElementLinks(workspace: Workspace, url: String) : C4
         withElementLinks.writeHeader(view, writer) { v, w -> super.writeHeader(v, w) }
     }
 
-    override fun writeElement(view: ModelView?, element: Element?, writer: IndentingWriter?) {
+    override fun writeElement(view: ModelView?, element: Element?,  writer: IndentingWriter?) {
         withElementLinks.writeElement(view, element, writer) { v, e, w -> super.writeElement(v, e, w) }
     }
 }
@@ -162,4 +163,5 @@ class StructurizrPlantUmlExporterWithElementLinks(workspace: Workspace, url: Str
     override fun writeElement(view: ModelView?, element: Element?, writer: IndentingWriter?) {
         withElementLinks.writeElement(view, element, writer) { v, e, w -> super.writeElement(v, e, w) }
     }
+
 }
